requirements.txt
asgiref==3.9.1
attrs==25.3.0
black==25.9.0
click==8.3.0
colorama==0.4.6
Django==5.2.5
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
model-bakery==1.20.5
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
pytokens==0.1.10
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
ruff==0.13.3
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
